Fix/Added small fixes after testing

+++ HealthCare/HealthCare.Domain/DTOs/PrescriptionDTO.cs
@@ -16,7 +16,7 @@ namespace HealthCare.Domain.DTOs
 
         public DateTime TakeAt { get; set; }
 
-        public int PerDay { get; set; }
+        public decimal PerDay { get; set; }
 
         public string MealCombination { get; set; }
     }
